drivers/card_reader/rtsx_usb: Set SSC based on card type
This commit reintroduces the SSC depth configuration previously
used in the rts5139 driver. This functionality was removed in commit
00d8521 ("staging: remove rts5139 driver code").

Reintroducing this mechanism allows SDR104 cards to work in certain
readers, such as the RTS5170, which with the current driver, will
corrupt the card as it tries to read from it.

Cc: Arnd Bergmann <arnd@arndb.de>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Sean Rhodes <sean@starlabs.systems>